Stability

+++ datanodes/core/node_listbox.py
@@ -58,7 +58,7 @@ class NodeListBox(QListWidget):
 
             drag = QDrag(self)
             drag.setMimeData(mimeData)
-            drag.setHotSpot(QPoint(pixmap.width() / 2, pixmap.height() / 2))
+            drag.setHotSpot(QPoint(int(pixmap.width() / 2), int(pixmap.height() / 2)))
             drag.setPixmap(pixmap)
 
             drag.exec_(Qt.MoveAction)
